Clean up polls-manager page with metaboxes.
The function display_polls_manager_page is now within polls-manager.php (it mostly just wraps what was there before). The file polls-manager.php is now conditionally loaded if the manager page is to be shown by WP, and it now includes a couple of functions to render the new metaboxes on that page.

Warning message for logs conditionally shown when current logging method uses cookies.

Remove redundant escape characters in messages: "Please check the \'Yes\' checkbox".